backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
networkx==3.2
openai==1.3.9
python-dotenv==1.0.0
pypdf==3.17.1
pydantic==2.5.0
pydantic-settings==2.1.0
